chore: migrate to formatUtils for better organization
