+++ islands/collection/CollectionDetailsHeader.tsx
@@ -1,5 +1,5 @@
 import { Collection, StampRow } from "globals";
-import { abbreviateAddress } from "$lib/utils/util.ts";
+import { abbreviateAddress } from "$lib/utils/formatUtils.ts";
 import { handleImageError } from "$lib/utils/imageUtils.ts";
 
 export const CollectionDetailsHeader = (
