+++ islands/stamp/details/StampRelatedGraph.tsx
@@ -1,7 +1,7 @@
 import dayjs from "$dayjs/";
 import relativeTime from "$dayjs/plugin/relativeTime";
 import StampHolders from "$components/stampDetails/StampHolders.tsx";
-import { abbreviateAddress } from "$lib/utils/util.ts";
+import { abbreviateAddress } from "$lib/utils/formatUtils.ts";
 
 dayjs.extend(relativeTime);
