+++ lib/utils/formatUtils.ts
@@ -1,0 +1,138 @@
+import { BigFloat } from "bigfloat/mod.ts";
+import { SATOSHIS_PER_BTC } from "$lib/utils/constants.ts";
+// Helper constant for satoshi conversion
+
+export function abbreviateAddress(
+  address?: string,
+  sliceLength: number = 4,
+): string {
+  if (!address) return "";
+  return `${address.slice(0, sliceLength)}...${address.slice(-sliceLength)}`;
+}
+
+export function formatBTCAmount(
+  btc: number,
+  options: {
+    includeSymbol?: boolean;
+    decimals?: number;
+    stripZeros?: boolean;
+  } = {},
+): string {
+  const {
+    includeSymbol = true,
+    decimals = 8,
+    stripZeros = true,
+  } = options;
+
+  const formatted = btc.toFixed(decimals);
+  const result = stripZeros ? stripTrailingZeros(formatted) : formatted;
+  return includeSymbol ? `${result} BTC` : result;
+}
+
+export function formatSatoshisToBTC(
+  satoshis: number,
+  options: {
+    includeSymbol?: boolean;
+    decimals?: number;
+    stripZeros?: boolean;
+  } = {},
+): string {
+  return formatBTCAmount(satoshis / SATOSHIS_PER_BTC, options);
+}
+
+export function formatNumber(value: number, decimals: number = 8): string {
+  return value.toLocaleString(undefined, {
+    minimumFractionDigits: decimals,
+    maximumFractionDigits: decimals,
+  });
+}
+
+export function formatDate(date: Date): string {
+  const locale = navigator.language || "en-US";
+  return date.toLocaleDateString(locale);
+}
+
+export function stripTrailingZeros(num: number | string): string {
+  const str = num.toString();
+  const parts = str.split(".");
+  if (parts.length === 1) {
+    return str; // No decimal point, return as is
+  }
+  const integerPart = parts[0];
+  const decimalPart = parts[1].replace(/0+$/, "");
+  return decimalPart.length > 0 ? `${integerPart}.${decimalPart}` : integerPart;
+}
+
+export function bigFloatToString(
+  value: BigFloat,
+  precision: number = 3,
+): string {
+  const stringValue = value.toString();
+  const [integerPart, fractionalPart] = stringValue.split(".");
+
+  if (!fractionalPart) {
+    return integerPart;
+  }
+
+  const roundedFractionalPart = fractionalPart.slice(0, precision);
+  const result = `${integerPart}.${roundedFractionalPart}`;
+
+  // Remove trailing zeros
+  return result.replace(/\.?0+$/, "");
+}
+
+export function formatSupplyValue(
+  supply: number | string | bigint,
+  divisible: boolean,
+): string {
+  if (typeof supply === "bigint") {
+    return divisible
+      ? (supply / 100000000n).toString()
+      : BigInt(supply).toString();
+  } else if (typeof supply === "string") {
+    supply = parseInt(supply);
+  }
+  return divisible ? (supply / 100000000).toFixed(2) : supply.toString();
+}
+
+export function isIntOr32ByteHex(value: string) {
+  const isInt = Number.isInteger(value) ||
+    (typeof value === "string" && Number.isInteger(Number(value)));
+
+  const is32ByteHex = typeof value === "string" &&
+    /^[0-9a-fA-F]{64}$/.test(value);
+
+  return isInt || is32ByteHex;
+}
+
+export function categorizeInput(
+  value: string | number,
+): "number" | "hex_string" | "none" {
+  if (
+    (typeof value === "string" && /^\d+$/.test(value)) ||
+    Number.isInteger(value)
+  ) {
+    return "number";
+  }
+
+  if (typeof value === "string" && /^[0-9a-fA-F]+$/.test(value)) {
+    return "hex_string";
+  }
+
+  return "none";
+}
+
+export function formatBigInt(value: bigint): string {
+  return value.toString();
+}
+
+export function bigIntSerializer(_key: string, value: unknown): unknown {
+  if (typeof value === "bigint") {
+    return value.toString();
+  }
+  return value;
+}
+
+export function jsonStringifyWithBigInt(obj: object): string {
+  return JSON.stringify(obj, bigIntSerializer);
+}
