+++ lib/utils/numberUtils.ts
@@ -1,0 +1,3 @@
+export function generateRandomNumber(min: number, max: number): number {
+  return Math.floor(Math.random() * (max - min + 1)) + min;
+}
