--- lib/utils/util.ts
@@ -1,255 +1,0 @@
-import { BigFloat } from "bigfloat/mod.ts";
-
-export const mimeTypes: { [key: string]: string } = {
-  "jpg": "image/jpeg",
-  "jpeg": "image/jpeg",
-  "png": "image/png",
-  "gif": "image/gif",
-  "svg": "image/svg+xml",
-  "tif": "image/tiff",
-  "jfif": "image/jpeg",
-  "jpe": "image/jpeg",
-  "pbm": "image/x-portable-bitmap",
-  "pgm": "image/x-portable-graymap",
-  "ppm": "image/x-portable-pixmap",
-  "pnm": "image/x-portable-anymap",
-  "apng": "image/apng",
-  "bmp": "image/bmp",
-  "webp": "image/webp",
-  "heif": "image/heif",
-  "heic": "image/heic",
-  "avif": "image/avif",
-  "ico": "image/x-icon",
-  "tiff": "image/tiff",
-  "svgz": "image/svg+xml",
-  "wmf": "image/wmf",
-  "emf": "image/emf",
-  "pcx": "image/pcx",
-  "djvu": "image/vnd.djvu",
-  "djv": "image/vnd.djvu",
-  "html": "text/html",
-  "txt": "text/plain",
-};
-
-// export const mimeTypesArray: [string, string][] = Object.entries(mimeTypes);
-export const mimeTypesArray: string[] = Object.values(mimeTypes);
-
-const SUPPORTED_UNICODE_FROM_INDEXER_CODE =
-  "\U0001f004\U0001f0cf\U0001f170\U0001f171\U0001f17e\U0001f17f\U0001f18e\U0001f191\U0001f192\U0001f193\U0001f194\U0001f195\U0001f196\U0001f197\U0001f198\U0001f199\U0001f19a\U0001f201\U0001f202\U0001f21a\U0001f22f\U0001f232\U0001f233\U0001f234\U0001f235\U0001f236\U0001f237\U0001f238\U0001f239\U0001f23a\U0001f250\U0001f251\U0001f300\U0001f301\U0001f302\U0001f303\U0001f304\U0001f305\U0001f306\U0001f307\U0001f308\U0001f309\U0001f30a\U0001f30b\U0001f30c\U0001f30d\U0001f30e\U0001f30f\U0001f310\U0001f311\U0001f312\U0001f313\U0001f314\U0001f315\U0001f316\U0001f317\U0001f318\U0001f319\U0001f31a\U0001f31b\U0001f31c\U0001f31d\U0001f31e\U0001f31f\U0001f320\U0001f321\U0001f324\U0001f325\U0001f642\U0001f326\U0001f327\U0001f328\U0001f329\U0001f32a\U0001f32b\U0001f32c\U0001f32d\U0001f32e\U0001f32f\U0001f330\U0001f331\U0001f332\U0001f333\U0001f334\U0001f335\U0001f336\U0001f337\U0001f338\U0001f339\U0001f33a\U0001f33b\U0001f33c\U0001f33d\U0001f33e\U0001f33f\U0001f340\U0001f341\U0001f342\U0001f343\U0001f344\U0001f345\U0001f346\U0001f347\U0001f348\U0001f349\U0001f34a\U0001f34b\U0001f34c\U0001f34d\U0001f34e\U0001f34f\U0001f350\U0001f351\U0001f352\U0001f353\U0001f354\U0001f355\U0001f356\U0001f357\U0001f358\U0001f359\U0001f35a\U0001f35b\U0001f35c\U0001f35d\U0001f35e\U0001f35f\U0001f360\U0001f361\U0001f362\U0001f363\U0001f364\U0001f365\U0001f366\U0001f367\U0001f368\U0001f369\U0001f36a\U0001f36b\U0001f36c\U0001f36d\U0001f36e\U0001f36f\U0001f370\U0001f371\U0001f372\U0001f373\U0001f374\U0001f375\U0001f376\U0001f377\U0001f378\U0001f379\U0001f37a\U0001f37b\U0001f37c\U0001f37d\U0001f37e\U0001f37f\U0001f380\U0001f381\U0001f382\U0001f383\U0001f384\U0001f385\U0001f386\U0001f387\U0001f388\U0001f389\U0001f38a\U0001f38b\U0001f38c\U0001f38d\U0001f38e\U0001f38f\U0001f390\U0001f391\U0001f392\U0001f393\U0001f396\U0001f397\U0001f399\U0001f39a\U0001f39b\U0001f39e\U0001f39f\U0001f3a0\U0001f3a1\U0001f3a2\U0001f3a3\U0001f3a4\U0001f3a5\U0001f3a6\U0001f3a7\U0001f3a8\U0001f3a9\U0001f3aa\U0001f3ab\U0001f3ac\U0001f3ad\U0001f3ae\U0001f3af\U0001f3b0\U0001f3b1\U0001f3b2\U0001f3b3\U0001f3b4\U0001f3b5\U0001f3b6\U0001f3b7\U0001f3b8\U0001f3b9\U0001f3ba\U0001f3bb\U0001f3bc\U0001f3bd\U0001f3be\U0001f3bf\U0001f3c0\U0001f3c1\U0001f3c2\U0001f3c3\U0001f3c4\U0001f3c5\U0001f3c6\U0001f3c7\U0001f3c8\U0001f3c9\U0001f3ca\U0001f3cb\U0001f3cc\U0001f3cd\U0001f3ce\U0001f3cf\U0001f3d0\U0001f3d1\U0001f3d2\U0001f3d3\U0001f3d4\U0001f3d5\U0001f3d6\U0001f3d7\U0001f3d8\U0001f3d9\U0001f3da\U0001f3db\U0001f3dc\U0001f3dd\U0001f3de\U0001f3df\U0001f3e0\U0001f3e1\U0001f3e2\U0001f3e3\U0001f3e4\U0001f3e5\U0001f3e6\U0001f3e7\U0001f3e8\U0001f3e9\U0001f3ea\U0001f3eb\U0001f3ec\U0001f3ed\U0001f3ee\U0001f3ef\U0001f3f0\U0001f3f3\U0001f3f4\U0001f3f5\U0001f3f7\U0001f3f8\U0001f3f9\U0001f3fa\U0001f400\U0001f401\U0001f402\U0001f403\U0001f404\U0001f405\U0001f406\U0001f407\U0001f408\U0001f409\U0001f40a\U0001f40b\U0001f40c\U0001f40d\U0001f40e\U0001f40f\U0001f410\U0001f411\U0001f412\U0001f413\U0001f414\U0001f415\U0001f416\U0001f417\U0001f418\U0001f419\U0001f41a\U0001f41b\U0001f41c\U0001f41d\U0001f41e\U0001f41f\U0001f420\U0001f421\U0001f422\U0001f423\U0001f424\U0001f425\U0001f426\U0001f427\U0001f428\U0001f429\U0001f42a\U0001f42b\U0001f42c\U0001f42d\U0001f42e\U0001f42f\U0001f430\U0001f431\U0001f432\U0001f433\U0001f434\U0001f435\U0001f436\U0001f437\U0001f438\U0001f439\U0001f43a\U0001f43b\U0001f43c\U0001f43d\U0001f43e\U0001f43f\U0001f440\U0001f441\U0001f442\U0001f443\U0001f444\U0001f445\U0001f446\U0001f447\U0001f448\U0001f449\U0001f44a\U0001f44b\U0001f44c\U0001f44d\U0001f44e\U0001f44f\U0001f450\U0001f451\U0001f452\U0001f453\U0001f454\U0001f455\U0001f456\U0001f457\U0001f458\U0001f459\U0001f45a\U0001f45b\U0001f45c\U0001f45d\U0001f45e\U0001f45f\U0001f460\U0001f461\U0001f462\U0001f463\U0001f464\U0001f465\U0001f466\U0001f467\U0001f468\U0001f469\U0001f46a\U0001f46b\U0001f46c\U0001f46d\U0001f46e\U0001f46f\U0001f470\U0001f471\U0001f472\U0001f473\U0001f474\U0001f475\U0001f476\U0001f477\U0001f478\U0001f479\U0001f47a\U0001f47b\U0001f47c\U0001f47d\U0001f47e\U0001f47f\U0001f480\U0001f481\U0001f482\U0001f483\U0001f484\U0001f485\U0001f486\U0001f487\U0001f488\U0001f489\U0001f48a\U0001f48b\U0001f48c\U0001f48d\U0001f48e\U0001f48f\U0001f490\U0001f491\U0001f492\U0001f493\U0001f494\U0001f495\U0001f496\U0001f497\U0001f498\U0001f499\U0001f49a\U0001f49b\U0001f49c\U0001f49d\U0001f49e\U0001f49f\U0001f4a0\U0001f4a1\U0001f4a2\U0001f4a3\U0001f4a4\U0001f4a5\U0001f4a6\U0001f4a7\U0001f4a8\U0001f4a9\U0001f4aa\U0001f4ab\U0001f4ac\U0001f4ad\U0001f4ae\U0001f4af\U0001f4b0\U0001f4b1\U0001f4b2\U0001f4b3\U0001f4b4\U0001f4b5\U0001f4b6\U0001f4b7\U0001f4b8\U0001f4b9\U0001f4ba\U0001f4bb\U0001f4bc\U0001f4bd\U0001f4be\U0001f4bf\U0001f4c0\U0001f4c1\U0001f4c2\U0001f4c3\U0001f4c4\U0001f4c5\U0001f4c6\U0001f4c7\U0001f4c8\U0001f4c9\U0001f4ca\U0001f4cb\U0001f4cc\U0001f4cd\U0001f4ce\U0001f4cf\U0001f4d0\U0001f4d1\U0001f4d2\U0001f4d3\U0001f4d4\U0001f4d5\U0001f4d6\U0001f4d7\U0001f4d8\U0001f4d9\U0001f4da\U0001f4db\U0001f4dc\U0001f4dd\U0001f4de\U0001f4df\U0001f4e0\U0001f4e1\U0001f4e2\U0001f4e3\U0001f4e4\U0001f4e5\U0001f4e6\U0001f4e7\U0001f4e8\U0001f4e9\U0001f4ea\U0001f4eb\U0001f4ec\U0001f4ed\U0001f4ee\U0001f4ef\U0001f4f0\U0001f4f1\U0001f4f2\U0001f4f3\U0001f4f4\U0001f4f5\U0001f4f6\U0001f4f7\U0001f4f8\U0001f4f9\U0001f4fa\U0001f4fb\U0001f4fc\U0001f4fd\U0001f4ff\U0001f500\U0001f501\U0001f502\U0001f503\U0001f504\U0001f505\U0001f506\U0001f507\U0001f508\U0001f509\U0001f50a\U0001f50b\U0001f50c\U0001f50d\U0001f50e\U0001f50f\U0001f510\U0001f511\U0001f512\U0001f513\U0001f514\U0001f515\U0001f516\U0001f517\U0001f518\U0001f519\U0001f51a\U0001f51b\U0001f51c\U0001f51d\U0001f51e\U0001f51f\U0001f520\U0001f521\U0001f522\U0001f523\U0001f524\U0001f525\U0001f526\U0001f527\U0001f528\U0001f529\U0001f52a\U0001f52b\U0001f52c\U0001f52d\U0001f52e\U0001f52f\U0001f530\U0001f531\U0001f532\U0001f533\U0001f534\U0001f535\U0001f536\U0001f537\U0001f538\U0001f539\U0001f53a\U0001f53b\U0001f53c\U0001f53d\U0001f549\U0001f54a\U0001f54b\U0001f54c\U0001f54d\U0001f54e\U0001f550\U0001f551\U0001f552\U0001f553\U0001f554\U0001f555\U0001f556\U0001f557\U0001f558\U0001f559\U0001f55a\U0001f55b\U0001f55c\U0001f55d\U0001f55e\U0001f55f\U0001f560\U0001f561\U0001f562\U0001f563\U0001f564\U0001f565\U0001f566\U0001f567\U0001f56f\U0001f570\U0001f573\U0001f574\U0001f575\U0001f576\U0001f577\U0001f578\U0001f579\U0001f57a\U0001f587\U0001f58a\U0001f58b\U0001f58c\U0001f58d\U0001f590\U0001f595\U0001f596\U0001f5a4\U0001f5a5\U0001f5a8\U0001f5b1\U0001f5b2\U0001f5bc\U0001f5c2\U0001f5c3\U0001f5c4\U0001f5d1\U0001f5d2\U0001f5d3\U0001f5dc\U0001f5dd\U0001f5de\U0001f5e1\U0001f5e3\U0001f5e8\U0001f5ef\U0001f5f3\U0001f5fa\U0001f5fb\U0001f5fc\U0001f5fd\U0001f5fe\U0001f5ff\U0001f600\U0001f601\U0001f602\U0001f603\U0001f604\U0001f605\U0001f606\U0001f607\U0001f608\U0001f609\U0001f60a\U0001f60b\U0001f60c\U0001f60d\U0001f60e\U0001f60f\U0001f610\U0001f611\U0001f612\U0001f613\U0001f614\U0001f615\U0001f616\U0001f617\U0001f618\U0001f619\U0001f61a\U0001f61b\U0001f61c\U0001f61d\U0001f61e\U0001f61f\U0001f620\U0001f621\U0001f622\U0001f623\U0001f624\U0001f625\U0001f626\U0001f627\U0001f628\U0001f629\U0001f62a\U0001f62b\U0001f62c\U0001f62d\U0001f62e\U0001f62f\U0001f630\U0001f631\U0001f632\U0001f633\U0001f634\U0001f635\U0001f636\U0001f637\U0001f638\U0001f639\U0001f63a\U0001f63b\U0001f63c\U0001f63d\U0001f63e\U0001f63f\U0001f640\U0001f641\U0001f642\U0001f643\U0001f644\U0001f645\U0001f646\U0001f647\U0001f648\U0001f649\U0001f64a\U0001f64b\U0001f64c\U0001f64d\U0001f64e\U0001f64f\U0001f680\U0001f681\U0001f682\U0001f683\U0001f684\U0001f685\U0001f686\U0001f687\U0001f688\U0001f689\U0001f68a\U0001f68b\U0001f68c\U0001f68d\U0001f68e\U0001f68f\U0001f690\U0001f691\U0001f692\U0001f693\U0001f694\U0001f695\U0001f696\U0001f697\U0001f698\U0001f699\U0001f69a\U0001f69b\U0001f69c\U0001f69d\U0001f69e\U0001f69f\U0001f6a0\U0001f6a1\U0001f6a2\U0001f6a3\U0001f6a4\U0001f6a5\U0001f6a6\U0001f6a7\U0001f6a8\U0001fab0\U0001fab1\U0001fab2\U0001fab3\U0001fab4\U0001fab5\U0001fab6\U0001fac0\U0001fac1\U0001fac2\U0001fad0\U0001fad1\U0001fad2\U0001fad3\U0001fad4\U0001fad5\U0001fad6";
-const formattedString =
-  SUPPORTED_UNICODE_FROM_INDEXER_CODE.replace(/U/g, "\\u{") + "}";
-const emojiCodePoints = formattedString.split("}{").map((emoji) =>
-  emoji.replace("\\u{", "").replace("}", "")
-);
-export const SUPPORTED_UNICODE = new Set(
-  emojiCodePoints.map((code) => parseInt(code, 16)),
-);
-
-const specialCases: { [key: string]: string } = {
-  "svg+xml": "svg",
-  "javascript": "js",
-  "plain": "txt",
-};
-
-export const getFileSuffixFromMime = (mimetype: string): string => {
-  if (!mimetype) return "unknown";
-  const suffix = mimetype.split("/")[1];
-  return specialCases[suffix] || suffix || "json";
-};
-
-/**
- * Returns a shortened version of the given address.
- * @param address - The address to be shortened.
- * @param number - The number of characters to keep at the beginning and end of the address. Default is 6.
- * @returns The shortened address.
- */
-export function abbreviateAddress(
-  address?: string,
-  sliceLength: number = 4,
-): string {
-  if (!address) return "";
-  return `${address.slice(0, sliceLength)}...${address.slice(-sliceLength)}`;
-}
-
-export function formatSatoshisToBTC(satoshis: number): string {
-  const btc = (satoshis / 100000000).toFixed(8);
-  return stripTrailingZeros(btc) + " BTC";
-}
-
-export function formatNumber(value: number, decimals: number = 8): string {
-  return value.toLocaleString(undefined, {
-    minimumFractionDigits: decimals,
-    maximumFractionDigits: decimals,
-  });
-}
-
-export function formatDate(date: Date): string {
-  const locale = navigator.language || "en-US";
-  return date.toLocaleDateString(locale);
-}
-
-export function stripTrailingZeros(num: number | string): string {
-  const str = num.toString();
-  const parts = str.split(".");
-  if (parts.length === 1) {
-    return str; // No decimal point, return as is
-  }
-  const integerPart = parts[0];
-  const decimalPart = parts[1].replace(/0+$/, "");
-  return decimalPart.length > 0 ? `${integerPart}.${decimalPart}` : integerPart;
-}
-
-export const getMimeType = (extension: string): string => {
-  const normalizedExt = extension.toLowerCase();
-
-  return mimeTypes[normalizedExt] || "application/octet-stream";
-};
-
-export const generateRandomNumber = (min: number, max: number) => {
-  return Math.floor(Math.random() * (max - min)) + min;
-};
-
-export const getSupply = (supply: number, divisible: boolean) => {
-  if (typeof supply === "bigint") {
-    return divisible
-      ? (supply / 100000000n).toString()
-      : BigInt(supply).toString();
-  } else if (typeof supply === "string") {
-    supply = parseInt(supply);
-  }
-  return divisible ? (supply / 100000000).toFixed(2) : supply;
-};
-
-export function isIntOr32ByteHex(value: string) {
-  // Check if the value is an integer or a numeric string that can be converted to an integer
-  const isInt = Number.isInteger(value) ||
-    (typeof value === "string" && Number.isInteger(Number(value)));
-
-  // Check if the value is a string representing a 32-byte hex (64 characters)
-  const is32ByteHex = typeof value === "string" &&
-    /^[0-9a-fA-F]{64}$/.test(value);
-
-  return isInt || is32ByteHex;
-}
-
-export function categorizeInput(
-  value: string | number,
-): "number" | "hex_string" | "none" {
-  // Check if the value is a strictly numeric string or a number
-  if (
-    (typeof value === "string" && /^\d+$/.test(value)) ||
-    Number.isInteger(value)
-  ) {
-    return "number";
-  }
-
-  // Check if the value is a string representing a hexadecimal string
-  if (typeof value === "string" && /^[0-9a-fA-F]+$/.test(value)) {
-    return "hex_string";
-  }
-
-  // If neither, return "none"
-  return "none";
-}
-
-export function paginate(total: number, page = 1, limit = 10) {
-  const totalPages = Math.ceil(total / limit);
-  return {
-    page,
-    limit,
-    totalPages,
-    total,
-  };
-}
-
-export function jsonStringifyBigInt(obj: object): string {
-  return JSON.stringify(obj, (_key: string, value: unknown) => {
-    if (typeof value === "bigint") {
-      return value.toString();
-    }
-    return value;
-  });
-}
-
-export function convertToEmoji(tick: string): string {
-  // Check if the string is already an emoji
-  if (
-    /^[\u{1F300}-\u{1F5FF}\u{1F600}-\u{1F64F}\u{1F680}-\u{1F6FF}\u{1F700}-\u{1F77F}\u{1F780}-\u{1F7FF}\u{1F800}-\u{1F8FF}\u{1F900}-\u{1F9FF}\u{1FA00}-\u{1FA6F}\u{1FA70}-\u{1FAFF}\u{2600}-\u{26FF}\u{2700}-\u{27BF}\u{2B50}-\u{2B55}\u{1F004}\u{1F0CF}]+$/u
-      .test(tick)
-  ) {
-    return tick;
-  }
-
-  // Convert Unicode code points to emojis
-  return tick.replace(/\\u([a-fA-F0-9]{8})/gi, (match, grp) => {
-    const codePoint = parseInt(grp, 16);
-    if (!isNaN(codePoint)) {
-      return String.fromCodePoint(codePoint);
-    } else {
-      return match;
-    }
-  });
-}
-
-export function convertEmojiToTick(str: string): string {
-  str = decodeURIComponent(str);
-  let result = "";
-  for (const char of str) {
-    if (char.match(/[\uD800-\uDBFF][\uDC00-\uDFFF]/)) {
-      result += "\\\\u" + char.codePointAt(0)!.toString(16).padStart(8, "0");
-    } else {
-      result += char;
-    }
-  }
-  return result;
-}
-
-export function bigFloatToString(
-  value: BigFloat,
-  precision: number = 3,
-): string {
-  const stringValue = value.toString();
-  const [integerPart, fractionalPart] = stringValue.split(".");
-
-  if (!fractionalPart) {
-    return integerPart;
-  }
-
-  const roundedFractionalPart = fractionalPart.slice(0, precision);
-  const result = `${integerPart}.${roundedFractionalPart}`;
-
-  // Remove trailing zeros
-  return result.replace(/\.?0+$/, "");
-}
-
-export function isValidSVG(svgContent: string): boolean {
-  const parser = new DOMParser();
-  const doc = parser.parseFromString(svgContent, "image/svg+xml");
-
-  // Check for parsing errors
-  const parserError = doc.querySelector("parsererror");
-  if (parserError) return false;
-
-  // Check for potentially dangerous elements
-  const dangerous = doc.querySelectorAll(
-    "foreignObject, use[href*='http'], image[href*='http'], a[href*='http']",
-  );
-  if (dangerous.length > 0) return false;
-
-  // Check all elements for event handlers and external references
-  const allElements = doc.getElementsByTagName("*");
-  for (const element of allElements) {
-    const attributes = element.attributes;
-    for (const attr of attributes) {
-      // Check for event handlers
-      if (attr.name.startsWith("on")) return false;
-
-      // Only block absolute URLs and data URIs
-      if (attr.value.match(/^(http|data):/i)) {
-        return false;
-      }
-    }
-  }
-
-  return true;
-}
