+++ routes/api/v2/src20/balance/[address]/[tick].ts
@@ -2,7 +2,7 @@ import { Handlers } from "$fresh/server.ts";
 import { AddressTickHandlerContext } from "globals";
 import { Src20Controller } from "$server/controller/src20Controller.ts";
 import { ResponseUtil } from "$lib/utils/responseUtil.ts";
-import { convertEmojiToTick } from "$lib/utils/util.ts";
+import { convertEmojiToTick } from "$lib/utils/emojiUtils.ts";
 
 export const handler: Handlers<AddressTickHandlerContext> = {
   async GET(req, ctx) {
