+++ routes/api/v2/src20/balance/snapshot/[tick].ts
@@ -1,7 +1,7 @@
 import { Handlers } from "$fresh/server.ts";
 import { AddressTickHandlerContext } from "globals";
 import { Src20Controller } from "$server/controller/src20Controller.ts";
-import { convertEmojiToTick } from "$lib/utils/util.ts";
+import { convertEmojiToTick } from "$lib/utils/emojiUtils.ts";
 import { ResponseUtil } from "$lib/utils/responseUtil.ts";
 import { getPaginationParams } from "$lib/utils/paginationUtils.ts";
