+++ routes/api/v2/src20/block/[block_index]/[tick].ts
@@ -1,6 +1,6 @@
 import { Handlers } from "$fresh/server.ts";
 import { Src20Controller } from "$server/controller/src20Controller.ts";
-import { convertEmojiToTick } from "$lib/utils/util.ts";
+import { convertEmojiToTick } from "$lib/utils/emojiUtils.ts";
 import { ResponseUtil } from "$lib/utils/responseUtil.ts";
 
 export const handler: Handlers = {
