+++ routes/api/v2/src20/tick/[tick]/index.ts
@@ -1,5 +1,5 @@
 import { Handlers } from "$fresh/server.ts";
-import { convertEmojiToTick, convertToEmoji } from "$lib/utils/util.ts";
+import { convertEmojiToTick, convertToEmoji } from "$lib/utils/emojiUtils.ts";
 import { PaginatedTickResponseBody } from "globals";
 import { ResponseUtil } from "$lib/utils/responseUtil.ts";
 import { Src20Controller } from "$server/controller/src20Controller.ts";
