+++ routes/api/v2/stamps/dispensers/index.ts
@@ -1,7 +1,6 @@
 import { Handlers } from "$fresh/server.ts";
-// import { paginate } from "$lib/utils/util.ts";
 import { ResponseUtil } from "$lib/utils/responseUtil.ts";
-// import { getPaginationParams } from "$lib/utils/paginationUtils.ts";
+// import { getPaginationParams, paginate } from "$lib/utils/paginationUtils.ts";
 // import { BlockController } from "$lib/controller/blockController.ts";
 // import { DispenserManager } from "$server/services/xcpService.ts";
