+++ routes/src20/[tick].tsx
@@ -1,7 +1,7 @@
 import { Handlers } from "$fresh/server.ts";
 import { SRC20TickHeader } from "$islands/src20/details/SRC20TickHeader.tsx";
 import { SRC20DetailsTab } from "$islands/src20/details/SRC20DetailsTab.tsx";
-import { convertEmojiToTick } from "$lib/utils/util.ts";
+import { convertEmojiToTick } from "$lib/utils/emojiUtils.ts";
 import { Src20Controller } from "$server/controller/src20Controller.ts";
 import { set_precision } from "bigfloat/mod.ts";
 import { SRC20RelatedGraph } from "$components/src20/SRC20RelatedGraph.tsx";
