+++ server/controller/src101Controller.ts
@@ -10,7 +10,8 @@ import {
 import { StampService } from "$server/services/stampService.ts";
 import { getBtcAddressInfo } from "$lib/utils/btc.ts";
 import { BlockService } from "$server/services/blockService.ts";
-import { convertToEmoji, paginate } from "$lib/utils/util.ts";
+import { convertToEmoji } from "$lib/utils/emojiUtils.ts";
+import { paginate } from "$lib/utils/paginationUtils.ts"
 
 export class Src101Controller{
   static async handleSrc101TXFromSRC101Table(
