+++ server/controller/src20Controller.ts
@@ -8,7 +8,7 @@ import {
 import { StampService } from "$server/services/stampService.ts";
 import { BTCAddressService } from "$server/services/btc/addressService.ts";
 import { BlockService } from "$server/services/blockService.ts";
-import { convertToEmoji } from "$lib/utils/util.ts";
+import { convertToEmoji } from "$lib/utils/emojiUtils.ts";
 import { SRC20MarketService } from "$server/services/src20/marketService.ts";
 import { MarketListingSummary } from "$types/index.d.ts";
 import { fetchBTCPriceInUSD } from "$lib/utils/btc.ts";
