+++ server/database/src101Repository.ts
@@ -1,6 +1,5 @@
 import { serverConfig } from "$server/config/config.ts";
-import { BigFloat } from "bigfloat/mod.ts";
-import { bigFloatToString } from "$lib/utils/util.ts";
+import { bigFloatToString } from "$lib/utils/formatUtils.ts";
 import {
   BIG_LIMIT,
   SRC101_OWNERS_TABLE,
